+++ CHANGELOG.md
@@ -2,7 +2,10 @@
 
 ## <a name="2.0.0"></a> 2.0.0 (2017-02-24)
 
-Cleaned up and simplified programming model.
+Cleaned up and simplified dependency management.
+
+### Features
+* **refer** Added **DependencyResolver**
 
 ### Breaking Changes
 * Refactored **refer** package. Removed IDescriptable and ILocateable interface. Made locator a mandatory requirement to place component into references.